requirements.txt
streamlit==1.17.0
torch==2.1.0
torchvision==0.15.0
Pillow==10.0.0
